requirements.txt
allure-python-commons==2.15.0
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.14.2
blinker==1.7.0
Brlapi==0.8.5
certifi==2025.10.5
chardet==5.2.0
click==8.1.6
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
duplicity==2.1.4
execnet==2.1.1
Faker==37.12.0
fasteners==0.18
h11==0.16.0
httplib2==0.20.4
idna==3.6
iniconfig==2.1.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
outcome==1.3.0.post0
packaging==25.0
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
pluggy==1.6.0
ptyprocess==0.7.0
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
PySocks==1.7.1
pytest==8.4.2
pytest-xdist==3.8.0
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
selenium==4.36.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
systemd-python==235
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.5.0
usb-creator==0.3.16
wadllib==1.3.6
websocket-client==1.9.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
